Continue integrating channel and spi infrastructure

+++ RTL/verilator/gps_tb/spi_host.h
@@ -4,8 +4,8 @@
 #include <stdint.h>
 
 void spi_init(Vgps_tb *const top);
-void spi_write(Vgps_tb *const top, uint8_t addr, uint8_t data);
-void spi_read(Vgps_tb *const top, uint8_t addr);
+void spi_write(Vgps_tb *const top, uint16_t addr, uint16_t data);
+void spi_read(Vgps_tb *const top, uint16_t addr);
 void spi_task(Vgps_tb *const top);
 bool spi_is_idle();
-uint8_t spi_get_read_data();
+uint16_t spi_get_read_data();
